requirements.txt
google-generativeai==0.8.3
tweepy==4.14.0
praw==7.7.1
requests==2.32.3
APScheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.3
responses==0.25.3
freezegun==1.5.1
